day 3

+++ Day2/BMICal.py
@@ -3,6 +3,6 @@ height = input("Enter height (m): ")
 #Enter weight
 weight = input("Enter weight (kg): ")
 h = float(height)
-w = int(weight)
+w = float(weight)
 BMI = w / (h*h)
 print("BMI = ", BMI)
